Added middleware to Route budget / changed guest home to login screen / changed user home to budget page

+++ database/migrations/2022_01_25_115020_create_expenses_table.php
@@ -16,7 +16,7 @@ class CreateExpensesTable extends Migration
         Schema::create('expenses', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->float('cost');
+            $table->decimal('cost', 10, 2);
             $table->foreignId('category_id')->constrained();
             $table->foreignId('budget_id')->constrained()->cascadeOnDelete();
             $table->foreignId('user_id')->constrained()->cascadeOnDelete();
